+++ R/layout.R
@@ -150,12 +150,13 @@ Layout <- ggproto("Layout", NULL,
 
   render_labels = function(self, labels, theme) {
     label_grobs <- lapply(names(labels), function(label) {
-      lapply(names(labels[[label]]), function(ary) {
-        if (is.null(labels[[label]][[ary]]) || is.waive(labels[[label]][[ary]])) return(zeroGrob())
+      lapply(c(1, 2), function(i) {
+        modify <- if (i == 2 && label == "y") ".right" else if (i == 1 && label == "x") ".top" else ""
+        if (is.null(labels[[label]][[i]]) || is.waive(labels[[label]][[i]])) return(zeroGrob())
         args <- list(
           theme = theme,
-          element = paste0("axis.title.", label),
-          label = labels[[label]][[ary]],
+          element = paste0("axis.title.", label, modify),
+          label = labels[[label]][[i]],
           expand_x = label == "y",
           expand_y = label == "x"
         )