+++ R/theme-defaults.r
@@ -81,17 +81,28 @@ theme_grey <- function(base_size = 11, base_family = "") {
     axis.line.y =        element_blank(),
     axis.text =          element_text(size = rel(0.8), colour = "grey30"),
     axis.text.x =        element_text(margin = margin(t = 0.8 * half_line / 2), vjust = 1),
+    axis.text.x.top =    element_text(margin = margin(b = 0.8 * half_line / 2), vjust = 0),
     axis.text.y =        element_text(margin = margin(r = 0.8 * half_line / 2), hjust = 1),
+    axis.text.y.right =  element_text(margin = margin(l = 0.8 * half_line / 2), hjust = 0),
     axis.ticks =         element_line(colour = "grey20"),
     axis.ticks.length =  unit(half_line / 2, "pt"),
     axis.title.x =       element_text(
                            margin = margin(t = half_line),
                            vjust = 1
                          ),
+    axis.title.x.top =   element_text(
+                           margin = margin(b = half_line),
+                           vjust = 0
+                         ),
     axis.title.y =       element_text(
                            angle = 90,
                            margin = margin(r = half_line),
-                           vjust = 1,
+                           vjust = 1
+                         ),
+    axis.title.y.right = element_text(
+                           angle = -90,
+                           margin = margin(l = half_line),
+                           vjust = 0
                          ),
 
     legend.background =  element_rect(colour = NA),